Rename h2 to connect
