+++ src/extension/serializer.ts
@@ -34,7 +34,7 @@ import {
 } from '../constants'
 import { ServerLifecycleIdentity } from '../utils/configuration'
 
-import { ParserService, createConnectClient, ConnectClient } from './grpc/parser/h2/client'
+import { ParserService, createConnectClient, ConnectClient } from './grpc/parser/connect/client'
 // ts bindings generated by protoc-gen-es
 import {
   RunmeIdentity,
@@ -47,7 +47,7 @@ import {
   DeserializeRequest,
   SerializeRequest,
   SerializeRequestOptions,
-} from './grpc/parser/h2/types'
+} from './grpc/parser/connect/types'
 import { type ReadyPromise } from './grpc/tcpClient'
 import Languages from './languages'
 import { PLATFORM_OS } from './constants'
